requirements.txt
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
xgboost>=1.4.0
requests>=2.26.0
python-dotenv>=0.19.0
web3>=5.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
optuna>=2.10.0
shap>=0.40.0
statsmodels>=0.13.0
scipy>=1.7.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
plotly>=5.0.0 